+++ lib/mcp_client/server_sse.rb
@@ -256,7 +256,7 @@ module MCPClient
     # @raise [MCPClient::Errors::ConnectionError] if server is disconnected
     def read_resource(uri)
       ensure_initialized
-      read_resource_with_cache(uri) { rpc_request('resources/read', { uri: uri }) }
+      read_resource_with_cache(uri) { |sent| rpc_request('resources/read', { uri: sent }) }
     rescue MCPClient::Errors::ServerError => e
       raise if e.protocol_error?
       raise resource_not_found_error(uri, e) if resource_not_found_response?(e)