+++ lib/mcp_client/server_streamable_http.rb
@@ -459,7 +459,7 @@ module MCPClient
     # @raise [MCPClient::Errors::ResourceReadError] if resource reading fails
     def read_resource(uri)
       ensure_connected
-      read_resource_with_cache(uri) { rpc_request('resources/read', { uri: uri }) }
+      read_resource_with_cache(uri) { |sent| rpc_request('resources/read', { uri: sent }) }
     rescue MCPClient::Errors::ServerError => e
       raise if e.protocol_error?
       raise resource_not_found_error(uri, e) if resource_not_found_response?(e)